+++ application/controllers/Main.php
@@ -20,10 +20,9 @@ class Main extends MY_Controller {
 	 */
 	public function index()
 	{
-		
 		$this->data['pagebody']='Homepage';
 		$this->data['pagetitle']='Homepage';
-		$this->data['PlayerInfo']=$this->StockModel->getPlayerInfo();
+		$this->data['PlayerInfo']=$this->PlayerModel->getPlayerInfo();
 		//$this->data['StockInfo']=$this->StockModel->getStockInfo();
 
 		$this->data['stockInfoArray']=$this->getStockArray();