+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.0.11 - 2016-02-18 - 10:44 PM
+### Added
+- Added functions to StockModel to retrieve stock transactions and stock movements.
+- Updated homepage dashboard to include links to the Stock History page when clicking on a stock.
+
+## 0.0.10 - 2016-02-18 - 2:00 PM
+### Added
+- Refactored player functions from StockModel to PlayerModel.
+
+## 0.0.9 - 2016-02-18 - 1:48 PM
+### Added
+- Moved friends panel to Homepage view.
+
 ## 0.0.9 - 2016-02-18 - 15:8 PM
 ### Added
 - Added stockhistory html
@@ -18,7 +31,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.0.6 - 2016-02-17 - 2:50 PM
 ### Added
-- Added login UI elements
+- integrated html/css into template
 
 ## 0.0.5 - 2016-02-16 - 1:41 pm
 ### Added